requirements.txt
# pip install --upgrade -r requirements.txt
fastapi~=0.115.8
itsdangerous~=2.2.0
mangum~=0.19.0
jinja2~=3.1.6
uvicorn~=0.34.0
pymysql~=1.1.1
starlette~=0.45.3
cryptography~=44.0.2
python-multipart~=0.0.20
bcrypt~=5.0.0
PyJWT~=2.12.1
python-dotenv~=1.2.2